vibe reafactor tool loading logic from input (url and cli args)
